requirements.txt
dashscope
sounddevice
keyboard
pyperclip
pystray
numpy
pillow
pyinstaller
